# entry/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Words api",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "localhost"
        },
        {
            "url": "https://api.vocasmart.pl",
            "description": "api.vocasmart.pl"
        }
    ],
    "paths": {
        "/api/flashcard/session": {
            "post": {
                "tags": [
                    "Flashcard"
                ],
                "summary": "Creates flashcard learning session for given category",
                "description": "Creates flashcard learning session for given category",
                "operationId": "flashcard.session.store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Requests\\Flashcard\\CreateSessionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Resources\\Flashcard\\SessionFlashcardsResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "firebase": []
                    }
                ]
            }
        },
        "/api/flashcard/session/{session_id}/rate-flashcards": {
            "post": {
                "tags": [
                    "Flashcard"
                ],
                "summary": "Rate session flashcards",
                "description": "Rate session flashcards and get next flashcards to learn",
                "operationId": "flashcard.session.rate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Requests\\Flashcard\\RateSessionFlashcardRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Resources\\Flashcard\\SessionFlashcardsResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "firebase": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "firebase": {
                "type": "http",
                "description": "Enter authorization token from firebase",
                "name": "Firebase Authorization",
                "scheme": "bearer"
            }
        }
    },
...
}